+++ page_test.go
@@ -6,17 +6,24 @@ import (
 	"github.com/gregoryv/workdir"
 )
 
+func Test_save_page_without_filename(t *testing.T) {
+	err := NewPage(Html()).SaveTo(".")
+	if err == nil {
+		t.Error("should fail")
+	}
+}
+
 func TestPage(t *testing.T) {
 	wd, err := workdir.TempDir()
 	if err != nil {
 		t.Fatal(err)
 	}
 	defer wd.RemoveAll()
-	page := NewPage("x.html", Html(Body()))
+	page := NewFile("x.html", Html(Body()))
 	page.SaveTo(string(wd))
 
 	wd.Chmod("x.html", 0000)
-	err = page.SaveTo(string(wd))
+	err = page.SaveAs(wd.Join("/x.html"))
 	if err == nil {
 		t.Error("should fail")
 	}